move qnn utils to subfolder
