+++ src/squlearn/qnn/__init__.py
@@ -3,7 +3,6 @@
 from .loss import ConstantLoss, ODELoss, ParameterRegularizationLoss, SquaredLoss, VarianceLoss
 from .qnnc import QNNClassifier
 from .qnnr import QNNRegressor
-from .training import get_variance_fac, get_lr_decay, ShotsFromRSTD
 
 __all__ = [
     "ConstantLoss",
@@ -13,7 +12,4 @@ __all__ = [
     "VarianceLoss",
     "QNNClassifier",
     "QNNRegressor",
-    "get_variance_fac",
-    "get_lr_decay",
-    "ShotsFromRSTD",
 ]
