+++ src/squlearn/qnn/util/__init__.py
@@ -1,0 +1,5 @@
+"""QNN Utilities."""
+
+from .training import get_variance_fac, get_lr_decay, ShotsFromRSTD
+
+__all__ = ["get_variance_fac", "get_lr_decay", "ShotsFromRSTD"]
