+++ src/squlearn/qnn/util/barren_plateau_analysis.py
@@ -1,9 +1,9 @@
 import numpy as np
 from sklearn.linear_model import LinearRegression
 
-from .lowlevel_qnn.lowlevel_qnn_base import LowLevelQNNBase
+from ..lowlevel_qnn.lowlevel_qnn_base import LowLevelQNNBase
 
-from ..util.data_preprocessing import adjust_parameters
+from ...util.data_preprocessing import adjust_parameters
 
 # TODO: This file needs a lot of revision. Not properly working with the current codebase.
